Add files via upload
Implemented a fully functioning GUI. Reorganized recording and playback features into accessible functions. Created a "Freeplay" function that acts the same as the "Recording" function but without recording any audio. The GUI includes three buttons for recording, audio playback, and freeplay, a countdown timer for each function, and a status bar that prompts the user what to do in every scenario.